+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## Unreleased
+
+The extension now requires at least [Visual Studio Code 1.42.0 _(January 2020)_](https://code.visualstudio.com/updates/v1_42).
+
+### 🚀 New Feature
+
+- Add support for [language-specific](https://code.visualstudio.com/docs/getstarted/settings#_languagespecific-editor-settings) custom toggles.
+
+### 🐛 Bug Fix
+
+- Prevent reloading the extension configuration when unrelated Visual Studio Code settings are modified.
+
 ## 0.1.1
 
 ### 🚀 New Feature